requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary
